Use native packager to generate runner scripts.
This Sbt plugin is the standard way to make executable projects in Scala
and gives a few more options, like the way to pass JVM arguments or system
properties, in addition to working on all platforms, including Windows.

The `universal:packageBin` task can produce an archive that can be used to
distribute the jars and runner scripts on all platforms.