Added a `stainlessIsEnabled` setting to the sbt-stainless plugin
Allows users to easily switch from using stainless to verify a module's
implementation to vanilla Scala compilation for actually producing binaries.

One downside of the current implementation is that the stainless library is
still compiled from sources even if `stainlessIsEnabled` is set to false. This
is due to the fact that the stainless-library binaries are not published.
However, if desiderable, it's easy to publish the stainless-library binaries and
overcome this limitation.