Replace reference to case object with a call to a constructor function.
This enables us to check the ADT invariant for case objects even if no one ever
directly references the case object. Previously, this was not the case and could
thus potentially lead to runtime errors.

Because the "constructor function" is marked with `@inline`, it will
disappear in the resulting program.