Check model for invalid ADT invariants in a more careful way.
Previously, the following testcase would correctly report an
invalid ADT invariant for `fail` without `--check-models`,
but when setting the latter option, it would be reported as
unknown because the model would to a runtime error,
ie. `Model leads to runtime error: ADT invariant violation for Foo(0)`.

```scala
object test {
  case class Foo(x: BigInt) {
    require(x != 0)
  }

  def fail = Foo(0)
}
```

Fixes #237