[WIP] Extern types
While we can already use `@extern` methods to hide
calls to methods defined by existing Scala libraries
which either make use of Scala features we do not support,
or are provided via the classpath, we currently have to way
to wrap the types exported by such libraries, because
the registry will complain that such types cannot be found
amongst the symbols.

This palliate to this problem, this PR adds support for annotating
types with the `@extern` annotation, which basically tells the
registry to effectively ignore such types from its dependency graph.

Attached is a simple testcase featuring a wrapper for
`scala.collection.concurrent.TrieMap`.