Implement ignored fields
While we can already use `@extern` methods to hide
calls to methods defined by existing Scala libraries
which either make use of Scala features we do not support,
or are provided via the classpath, we currently have no way
of wrapping the types exported by such libraries, because
the registry will complain that such types cannot be found
amongst the symbols.

This palliate to this problem, this PR adds support for annotating
fields with the `@ignore` annotation, which effectively replace
the type of the field by `Int`. It is safe and sound to do so,
because such fields are not meant to be accessed via non-`@extern`
methods, and we will thus never actually care about their real type.

Attached is a simple testcase featuring a wrapper for
`scala.collection.concurrent.TrieMap`.