Fix soundness issue in MethodLifting.
In a method's dispatcher, we now constrain forwarded arguments to
have the expected type, rather than just casting them to said type.

Not doing leads in some cases to invalid models where some argument's
model is of a totally unrelated type.

It is sound to use `assume` for this, as this is a constraint that
is guaranteed by both Scala's and our type system.