Help solver deduce that reified type args do not influence evaluation
If function has gained new parameters for reifed types during
TypeEncoding, we create a dummy version of it without any
type arguments as a witness that type arguments do not
influence the result of the computation of its invocation.

We do this by duplicating the encoded function into
another function without the reified type parameters
and an empty body.

The post of the encoded function is then modified to assert
that its result is equals to the result of this new function,
when applied to (non-type args) parameters.

This allows the solver to deduce that the reified type parameters
themselves do not influence the result of a call
to the original function, something that is needed eg.,
in the presence of GADTs where equivalent calls to the same method
via both the top-level dispatcher, and a concrete implementation
can carry different reified type parameters.