Provide option for batched compilation and post-compilation filtering
This commit provides the option to generate a Symbols object containing
the whole input program, and filter it for every invocation of the
Component.apply method as needed. This filtering is implemented with a
simple but sound dependency analysis on the level of stainless trees.

In more detail, this commit introduces the following changes:

- A new option --batched (off by default) provides the option discussed
  in the introduction.
- BatchedCallBack implements CallBack when --batched is on.
- The Report classes defined in StainlessCallBack have been refactored
  into a separate trait.
- DependenciesFinder gets renamed to XLangDependenciesFinder
- A new class, DependenciesFinder is introduced which operates on
  Stainless trees.
- The abstract Component.apply is renamed to Component.execute. The
  concrete Component.apply provides the option to filter trees with a
  DependenciesFinder. This option is exercised by BatchedCallBack.