Allow setting default values for options via config file
The file must be named `stainless.conf` and be a valid HOCON file.

It will be searched for recursively starting from the current
directory and walking up the directory hierarchy.

So, if one runs stainless from `/a/b/c`, and there is a config
file in any of `c`, `b` or `a`, the first of those is going to be
loaded.

For example, the config file could contain the following:

  vc-cache = false
  debug = [verification, trees]
  timeout = 5
  check-models = true
  print-ids = true

That will translate to the following command line options:

  --vc-cache=false --debug=verification,trees --timeout=5 --print-ids