Fix bug in VC generation within the typechecker
Term variables which were bound to a value were
incorrectly treated as truth values instead of
let equalities, which meant that those variables
were not turned into lets, but rather equalities which
could be disproven by the solver.

For example:

val a = 1
val b = 2
a == b

would yield

    (a == 1 && b == 2) ==> (a == b)

instead of

    val a = 1
    val b = 2
    ==>
    a == b

And thus in the first case, the solver was at liberty of
picking a = 1 and b = 1, which would make the VC valid.